+++ CmsWeb/Areas/Public/Models/MobileAPI/MobileGivingHistory.cs
@@ -5,14 +5,14 @@ namespace CmsWeb.MobileAPI
 	public class MobileGivingHistory
 	{
 		public int thisYear = 0;
-		public decimal yearToDateTotal = 0;
+		public int yearToDateTotal = 0;
 
 		public int lastYear = 0;
-		public decimal lastYearTotal = 0;
+		public int lastYearTotal = 0;
 
 		public List<MobileGivingEntry> entries = new List<MobileGivingEntry>();
 
-		public void setLastYearTotal( int lastYear, decimal lastYearTotal )
+		public void setLastYearTotal( int lastYear, int lastYearTotal )
 		{
 			this.lastYear = lastYear;
 			this.lastYearTotal = lastYearTotal;
